project 1/requirements.txt
Flask==3.0.0
deepface==0.0.79
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
flask-cors==4.0.0
